more NCC recommendations

+++ contracts/CstLedger.sol
@@ -4,14 +4,14 @@ import "./administratable.sol";
 import "zeppelin-solidity/contracts/math/SafeMath.sol";
 
 contract ITokenLedger {
-  function totalTokens() constant returns (uint);
-  function totalInCirculation() constant returns (uint);
-  function tokensAvailable() constant returns (uint);
-  function balanceOf(address account) constant returns (uint);
-  function mintTokens(uint amount);
-  function transfer(address sender, address reciever, uint amount);
-  function creditAccount(address account, uint amount);
-  function debitAccount(address account, uint amount);
+  function totalTokens() constant returns (uint256);
+  function totalInCirculation() constant returns (uint256);
+  function tokensAvailable() constant returns (uint256);
+  function balanceOf(address account) constant returns (uint256);
+  function mintTokens(uint256 amount);
+  function transfer(address sender, address reciever, uint256 amount);
+  function creditAccount(address account, uint256 amount);
+  function debitAccount(address account, uint256 amount);
   function addAdmin(address admin);
   function removeAdmin(address admin);
 }
@@ -20,30 +20,30 @@ contract CstLedger is ITokenLedger, administratable {
 
   using SafeMath for uint256;
 
-  uint public _totalInCirculation;
-  uint public _totalTokens;
-  mapping (address => uint) public _balanceOf;
-  uint public ledgerCount;
-  mapping (uint => address) public accountForIndex;
+  uint256 public _totalInCirculation;
+  uint256 public _totalTokens;
+  mapping (address => uint256) public _balanceOf;
+  uint256 public ledgerCount;
+  mapping (uint256 => address) public accountForIndex;
   mapping (address => bool) public accounts;
 
-  function totalTokens() constant returns (uint) {
+  function totalTokens() constant returns (uint256) {
     return _totalTokens;
   }
 
-  function totalInCirculation() constant returns (uint) {
+  function totalInCirculation() constant returns (uint256) {
     return _totalInCirculation;
   }
 
-  function tokensAvailable() constant returns (uint) {
+  function tokensAvailable() constant returns (uint256) {
     return _totalTokens.sub(_totalInCirculation);
   }
 
-  function balanceOf(address account) constant returns (uint) {
+  function balanceOf(address account) constant returns (uint256) {
     return _balanceOf[account];
   }
 
-  function mintTokens(uint amount) onlyAdmins {
+  function mintTokens(uint256 amount) onlyAdmins {
     _totalTokens = _totalTokens.add(amount);
   }
 
@@ -55,7 +55,7 @@ contract CstLedger is ITokenLedger, administratable {
     }
   }
 
-  function transfer(address sender, address recipient, uint amount) onlyAdmins {
+  function transfer(address sender, address recipient, uint256 amount) onlyAdmins {
     require(_balanceOf[sender] >= amount);
 
     _balanceOf[sender] = _balanceOf[sender].sub(amount);
@@ -63,14 +63,14 @@ contract CstLedger is ITokenLedger, administratable {
     makeAccountIterable(recipient);
   }
 
-  function creditAccount(address account, uint amount) onlyAdmins { // remove tokens
+  function creditAccount(address account, uint256 amount) onlyAdmins { // remove tokens
     require(_balanceOf[account] >= amount);
 
     _totalInCirculation = _totalInCirculation.sub(amount);
     _balanceOf[account] = _balanceOf[account].sub(amount);
   }
 
-  function debitAccount(address account, uint amount) onlyAdmins { // add tokens
+  function debitAccount(address account, uint256 amount) onlyAdmins { // add tokens
     _totalInCirculation = _totalInCirculation.add(amount);
     _balanceOf[account] = _balanceOf[account].add(amount);
     makeAccountIterable(account);
