+++ contracts/CstLibrary.sol
@@ -18,35 +18,35 @@ library CstLibrary {
     ExternalStorage(_storage).setBytes32Value("cstTokenSymbol", tokenName);
   }
 
-  function getBuyPrice(address _storage) constant returns(uint) {
+  function getBuyPrice(address _storage) constant returns(uint256) {
     return ExternalStorage(_storage).getUIntValue("cstBuyPrice");
   }
 
-  function setBuyPrice(address _storage, uint value) {
+  function setBuyPrice(address _storage, uint256 value) {
     ExternalStorage(_storage).setUIntValue("cstBuyPrice", value);
   }
 
-  function getSellPrice(address _storage) constant returns(uint) {
+  function getSellPrice(address _storage) constant returns(uint256) {
     return ExternalStorage(_storage).getUIntValue("cstSellPrice");
   }
 
-  function setSellPrice(address _storage, uint value) {
+  function setSellPrice(address _storage, uint256 value) {
     ExternalStorage(_storage).setUIntValue("cstSellPrice", value);
   }
 
-  function getSellCap(address _storage) constant returns(uint) {
+  function getSellCap(address _storage) constant returns(uint256) {
     return ExternalStorage(_storage).getUIntValue("cstSellCap");
   }
 
-  function setSellCap(address _storage, uint value) {
+  function setSellCap(address _storage, uint256 value) {
     ExternalStorage(_storage).setUIntValue("cstSellCap", value);
   }
 
-  function getMinimumBalance(address _storage) constant returns(uint) {
+  function getMinimumBalance(address _storage) constant returns(uint256) {
     return ExternalStorage(_storage).getUIntValue("cstMinimumBalance");
   }
 
-  function setMinimumBalance(address _storage, uint value) {
+  function setMinimumBalance(address _storage, uint256 value) {
     ExternalStorage(_storage).setUIntValue("cstMinimumBalance", value);
   }
 
@@ -58,11 +58,11 @@ library CstLibrary {
     ExternalStorage(_storage).setAddressValue("cstFoundation", value);
   }
 
-  function getAllowance(address _storage, address account, address spender) constant returns (uint) {
+  function getAllowance(address _storage, address account, address spender) constant returns (uint256) {
     return ExternalStorage(_storage).getMultiLedgerValue("cstAllowance", account, spender);
   }
 
-  function setAllowance(address _storage, address account, address spender, uint allowance) {
+  function setAllowance(address _storage, address account, address spender, uint256 allowance) {
     ExternalStorage(_storage).setMultiLedgerValue("cstAllowance", account, spender, allowance);
   }
